Add column to CreateBooks table in order for delete request logic to function correctly

+++ app/controllers/books_controller.rb
@@ -8,6 +8,8 @@ class BooksController < ApplicationController
             redirect to "/"
           end
     end
+      #CREATE --> NEW ACTION, renders new view
+
 
     get '/books' do 
         @books = Book.all 
@@ -18,6 +20,9 @@ class BooksController < ApplicationController
             redirect to "/users"
           end       
     end
+
+      #CREATE --->create action, handles form data, then redirects
+
 
     post '/books' do
         #binding.pry
@@ -31,6 +36,9 @@ class BooksController < ApplicationController
       end
     end 
 
+        #READ: index action, renders an index view 
+
+
      get '/books/:id' do 
         @book = Book.find_by_id(params[:id])
         if session[:user] == @book.user
@@ -39,22 +47,34 @@ class BooksController < ApplicationController
         redirect '/books' #to routes 
         end 
     end 
+  #READ: show action, renders a show view 
+
 
      get '/books/:id/edit' do
         @book = Book.find_by_id(params[:id])
         erb :'books/edit'
      end
+    #UPDATE: edit action, renders edit view
+
 
      patch '/books/:id' do
         book = Book.find_by_id(params[:id])
         book.update(book_title: params["Book Title"], author: params["Author"], book_genre: params["Book Genre"], number_of_pages: params["Number of Pages"], start: params["start"], end: params["end"], time_one: params["time_1"], time_two: params["time_2"], read: params["read"])
         redirect "/books/#{book.id}"
       end
+            #UPDATE action, handles form data, then redirects 
 
-      delete '/books/:id/delete' do
+
+      delete '/books/:id' do
         book = Book.find_by_id(params[:id])
-        book.destroy
+        binding.pry
+        if session[:user_id] !=book.user_id
+          redirect '/books'
+        else 
+          book.destroy
         redirect '/books'
-
+        end 
       end
+          #DELETE: deletes action, deletes resource, then redirects --> triggered by button in the show and or index view
+
 end
